1.2.8 is basically broken due to untested changes on a uncovered part of the code
Untested API calls regarding Translation module and lack of use of the `logging` library
concealed insidious errors JSON format

+++ api/servicemanager/__init__.py
@@ -20,9 +20,10 @@ class ProcessManager:
     spawned_backend_process = None
 
     def __del__(self):
-        self.spawned_backend_process.terminate()
-        path = '/tmp/%s.pid' % self.program_name
-        os.system('rm %s' % path)  # Process has died, pid file is irrelevant
+        if self.spawned_backend_process:
+            self.spawned_backend_process.terminate()
+            path = '/tmp/%s.pid' % self.program_name
+            os.system('rm %s' % path)  # Process has died, pid file is irrelevant
 
     def get_specific_arguments(self) -> List[str]:
         """
@@ -78,24 +79,26 @@ class ServiceManager(ProcessManager):
         """
         self.backend_address = addr
 
+    def get_url_head(self):
+        return '%s://%s:%d' % (self.scheme, self.backend_address, self.port)
 
     # Low-level functions to use with high-level functions
-    @retry_on_fail([Exception])
+    @retry_on_fail([Exception], 5, .5)
     def get(self, route, **kwargs):
         route = '%s://%s:%d/%s' % (self.scheme, self.backend_address, self.port, route)
         return requests.get(route, **kwargs)
 
-    @retry_on_fail([Exception])
+    @retry_on_fail([Exception], 5, .5)
     def post(self, route, **kwargs):
         route = '%s://%s:%d/%s' % (self.scheme, self.backend_address, self.port, route)
         return requests.post(route, **kwargs)
 
-    @retry_on_fail([Exception])
+    @retry_on_fail([Exception], 5, .5)
     def put(self, route, **kwargs):
         route = '%s://%s:%d/%s' % (self.scheme, self.backend_address, self.port, route)
         return requests.put(route, **kwargs)
 
-    @retry_on_fail([Exception])
+    @retry_on_fail([Exception], 5, .5)
     def delete(self, route, **kwargs):
         route = '%s://%s:%d/%s' % (self.scheme, self.backend_address, self.port, route)
         return requests.delete(route, **kwargs)
@@ -118,7 +121,13 @@ class LanguageServiceManager(ServiceManager):
     # high-level function
     def get_language(self, language_code) -> Response:
         result = self.get('language/' + language_code)
-        if result is not None:
-            return result
-        else:
-            raise Exception("Language not found")
+
+        for i in range(10):
+            if result is not None:
+                return result
+            else:
+                print(result)
+                time.sleep(.5)
+                result = self.get('language/' + language_code)
+
+        raise Exception("Language not found: " + language_code)
